==> NMps_book/01_downloading_data.rmd <==
# Downloading data 

```{r, message = FALSE}
library(R.utils)
library(tidyverse)
```

```{r}

urls = c("https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSM3067194&format=file&file=GSM3067194%5F18hpf%2Ecsv%2Egz", "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSM3067194&format=file&file=GSM3067194%5F18hpf%5FclustID%2Etxt%2Egz","https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE112294&format=file&file=GSE112294%5FClusterNames%2Ecsv%2Egz")

destination = c("./hpf18_count_matrix.csv.gz", "./hpf18_clustID.txt.gz", "./hpf18_clustNames.csv.gz" )

for (i in 1:length(urls)){
  download.file(urls[i], destination[i], mode = "wb") 
  gunzip(destination[i])
}

```
